server/requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
filelock==3.6.0
Flask==2.0.3
Flask-Cors==3.0.10
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.1
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.1.0
numpy==1.21.5
packaging==21.3
pycodestyle==2.8.0
pycryptodome==3.14.1
pyparsing==3.0.7
PyYAML==6.0
regex==2022.1.18
requests==2.27.1
rocksdb==0.7.0
sacremoses==0.0.47
six==1.16.0
tokenizers==0.11.5
toml==0.10.2
torch==1.10.2
tqdm==4.62.3
transformers==4.16.2
typing-extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
